Added ability to delete a variable by assignement.
Mainly intended for tasks to remove variable to ensure that only the
variable is set for a specific task via task-specific variable:

    @env.task(VAR=Delete())
    def task():
        env["VAR"] = Default("something")

Will result in VAR, within the task, being the value "something" unless
set to some other variable from the command line as a task variable:

    mrbavii-taskrun task:VAR=something_else

This works because task-specific variables defined in the file are first
applied, followed by task-specific variable from the command line.  As a
result, when the task is called, VAR is first deleted, and then if VAR
is a task-specific variable on the command line then set again to that
variable.  Otherwise, the default value will be applied.