Simplified completion script.
Previous completion script only extracted some variables.  Now the
entire command is used with the "--list" option.  This allows for
variables set on the command line to be passed taskrun so in case some
task files are conditionally included based on a variable name, it
should list the task names.